Cleanup DataBaseDriver
