--- selcukes-commons/src/main/java/io/github/selcukes/commons/db/DataBaseType.java
@@ -1,26 +1,0 @@
-/*
- *  Copyright (c) Ramesh Babu Prudhvi.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package io.github.selcukes.commons.db;
-
-public enum DataBaseType {
-    MY_SQL,
-    SQL_SERVER,
-    POST_GRE_SQL,
-    ORACLE,
-    ORACLE_SERVICE_NAME,
-    IBM_DB2
-}
